requirements.txt
Flask==3.1.1
python-dotenv==1.1.1
requests==2.32.4
oauthlib==3.1.1
requests-oauthlib==2.0.0
Pillow==11.3.0
python-slugify==8.0.4
Werkzeug==3.1.3